Move extent and mask_by_extent to base class
